# front-office/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/backoffice/catalog/categories/{categoryId}/modifier-groups": {
      "put": {
        "operationId": "CatalogCategoryModifiersController_replace",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCategoryModifierGroupsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryModifierGroupResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiHttpErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Заменить группы добавок категории",
        "tags": [
          "backoffice"
        ]
      }
    },
...
  },
  "info": {
    "title": "Expressa API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "expressa_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "expressa_refresh"
      }
    },
    "schemas": {
...
      "ApiHttpErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ]
      },
...
      "ApiValidationFieldDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "reason"
        ]
      },
      "ApiValidationErrorDetailsDto": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiValidationFieldDto"
            }
          }
        },
        "required": [
          "fields"
        ]
      },
      "ApiValidationErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/ApiValidationErrorDetailsDto"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ]
      },
...
      "ReplaceCategoryModifierGroupsDto": {
        "type": "object",
        "properties": {
          "groupIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "groupIds"
        ]
      },
      "CategoryModifierGroupResponseDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "categoryId",
          "groupId",
          "sortOrder"
        ]
      },
...
    }
  }
}